Error teste

+++ src/main/java/com/apiptow/dto/SkillDTO.java
@@ -1,11 +1,6 @@
 package com.apiptow.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
-
-import javax.persistence.OneToMany;
-
-import com.apiptow.entity.Person;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -21,7 +16,7 @@ public class SkillDTO {
 	private Long id;
 
 	private String name;
-	private BigDecimal salary;
+	private Double salary;
     private List<PersonDTO> peoples;
